ton of debugging

+++ src/flags.zig
@@ -1,4 +1,4 @@
-pub const Flags = enum(u32) {
+pub const Flags = enum(u8) {
     FL_POS = 1 << 0, // P
     FL_ZRO = 1 << 1, // Z
     FL_NEG = 1 << 2, // N
